Initial Commit
This includes the basic framework of a Python Lambda boilerplate
based largely off the python-lambda package with several features
added. Basic commands are run out of a makefile and includes basic
logging and error helpers.

This code should output a simple hello, world statement on success.

There are several features to be implemented, most importantly the
ability to set an event source mapping and build/deploy time and
run the lambda automatically with that.