SFR-234 Added working version of boilerplate lambda with event sources
This boilerplate lambda now has the ability to fully deploy to AWS
and includes the ability to add event source mappings for different
environments.

It also fully manages environment variables and allows users to set
different variables depending on the deployment environment. The
only remaining tasks are to create a repository and test the
integration with Travis to see if deployments will work from there