Add setup.cfg to gitignore and handle deploys with no sources
It may be necessary for some users to configure their installs with
a setup.cfg file to configure their virtualenvs, but including by
default can cause issues with travis, and also other users.

Additionally this commit fixes a bug that causes deployments to fail
if no event sources exist for a deployment. Since this is a valid
deployment option, it is necessary to handle this case. An additional
test was added to cover this conditional as well