Replace event source mapping with dict comprehension
Previously the event source mapping objects were being built with
explicit dictionary keys, which made it difficult to manage different
event source types. Switching to a dict comprehension allows
different sources to be easily handled.

The only drawback to this is that it relies on the user to provide a
properly formatted event source mapping. However, if a faulty event
source mapping is used the deployment process will fail with a fairly
descriptive error messsage that should allow the error to be easily
fixed.