Replace `ChainMap` with faster `dict` merging
Implenting the merging of two configuration dictionaries with
`ChainMap` proved to be unnecessary/slower. This has been removed,
with the replacement tested.

Revealed in this testing is that neither the merge nor `ChainMap`
will update nested `dict`s, merely replace them wholesale. This
means that configuration settings, particularly environment
variables must be duplicated wholesale across environments, not
only those that change. This should be an area of improvement
in the future.