Refactor `lambdaRun` for better maintainability
This refactors the `regex` based structure of the file that
managed the specific type of invocation being made to run the
lambda function.

This is replaced with an explicit mapping between the valid
execution commands and the methods that they invoke. In the case
that a command is received that is not found and error method
is called, which raises an `InvalidExectutionType` error.

This extracts a large amount of functionality from a single method
into several smaller methods, allowing for easier tracing of issues
and clearer visualisation of the path through this module.