Add env var decrypt method
It is necessary to have a means of decrypting KMS-encrypted environment variables. This provides a secure means of storing senstive information in Github, which makes it much easier to create a CI/CD pipeline.

The helper method here simply attempts to read the provided environment variable as an encrypted string, if it fails it assumes that it has received a non-encoded variable and returns it unaltered.