Fix config loading errors
Two bugs were introduced in the recent release of this function. First, the order of loading default and environment-specific environment variables was reversed, preventing certain variables from being loaded on a per-environment basis. This fixes by simply reversing the dictionary unpacking position. Second, the `configLines` variable has been removed, which was not reflected in a certain line where it was expected to be part of a tuple unpacking. This was simply removed. Both these changes necesitated updates to the test suite.