Ensure EventSourceArn is removed from update event source
The event source mapping update event is called when an event source is already in place with a function. If it is called with `EventSourceArn` as a parameter it will fail as it is not expecting this value. There where instances where this was mistakenly being passed to the method and causing errors in the automated build process.

This enforces its removal with `pop` and ensures that this behavior executes consistently with a stricter implemntation of a unit test